+++ core/Tracker/Config/ThirdPartyCookies.php
@@ -67,12 +67,12 @@ class ThirdPartyCookies implements
         return $currentValue === $policyValues[$policy];
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate('General_ThirdPartyCookieComplianceDescription');
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate('General_ThirdPartyCookieComplianceImpact');
     }