+++ plugins/DevicesDetection/Settings/DeviceModelDetectionDisabled.php
@@ -13,12 +13,12 @@ class DeviceModelDetectionDisabled extends CompliancePolicyEnforcedSetting
         return Piwik::translate('DevicesDetection_DeviceModelDetectionDisabled');
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate('DevicesDetection_DeviceModelDetectionComplianceDescription');
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate('DevicesDetection_DeviceModelDetectionComplianceImpact');
     }