+++ plugins/DevicesDetection/Settings/OnlyMajorVersions.php
@@ -78,12 +78,12 @@ class OnlyMajorVersions implements
         return $this->value;
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate("DevicesDetection_OnlyMajorVersionsComplianceDescription");
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate("DevicesDetection_OnlyMajorVersionsComplianceImpact");
     }