+++ plugins/Live/Settings/VisitorLogDisabled.php
@@ -92,12 +92,12 @@ class VisitorLogDisabled implements MeasurableSettingInterface, PolicyComparison
         return Piwik::translate('Live_VisitorLogPolicyComplianceTitle');
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate('Live_VisitorLogPolicyComplianceDescription');
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate('Live_VisitorLogPolicyComplianceImpact');
     }