+++ plugins/PrivacyManager/Settings/CampaignParameterValuesMasked.php
@@ -119,12 +119,12 @@ class CampaignParameterValuesMasked implements
         return $this->value;
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate('PrivacyManager_CampaignParameterValuesMaskedComplianceDescription');
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate('PrivacyManager_CampaignParameterValuesMaskedComplianceImpact');
     }