+++ plugins/PrivacyManager/Settings/DataRoundingEnabled.php
@@ -21,12 +21,12 @@ class DataRoundingEnabled extends CompliancePolicyEnforcedSetting
         return Piwik::translate('PrivacyManager_SegmentedDataRoundingSettingTitle');
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate('PrivacyManager_SegmentedDataRoundingComplianceDescription');
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate('PrivacyManager_SegmentedDataRoundingComplianceImpact');
     }