+++ plugins/SegmentEditor/Settings/LimitSegments.php
@@ -25,12 +25,12 @@ class LimitSegments extends CompliancePolicyEnforcedSetting
         ];
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate("SegmentEditor_LimitSegmentsComplianceDescription");
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate("SegmentEditor_LimitSegmentsComplianceImpact");
     }