+++ plugins/SitesManager/Settings/FilterPIIParameters.php
@@ -63,12 +63,12 @@ class FilterPIIParameters implements
         ];
     }
 
-    public static function getComplianceRequirementNote(?int $idSite = null): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return Piwik::translate('SitesManager_FilterPiiParametersComplianceDescription');
     }
 
-    public static function getComplianceImpactNote(?int $idSite = null): string
+    public static function getImpact(?int $idSite = null): string
     {
         return Piwik::translate('SitesManager_FilterPiiParametersComplianceImpact');
     }