+++ tests/PHPUnit/Framework/Mock/Settings/FakePolicySetting.php
@@ -62,14 +62,19 @@ class FakePolicySetting implements PolicyComparisonInterface, SettingValueInterf
         return false;
     }
 
-    public static function getWhatItDoes(): string
+    public static function getWhatItDoes(?int $idSite = null): string
     {
         return 'fake policy setting what it does';
     }
 
-    public static function getImpact(): string
+    public static function getImpact(?int $idSite = null): string
     {
         return 'fake policy setting impact';
+    }
+
+    public static function getComplianceTitle(?int $idSite = null): string
+    {
+        return static::getTitle();
     }
 
     public static function getInstance(?int $idSite = null)